+++ R/onAttach.R
@@ -4,8 +4,9 @@
   if ( interactive() ) {
     packageStartupMessage( 'Welcome to msmtools v', as.character( packageVersion( "msmtools" ) ),
                            '! An easy way to build augmented longitudinal datasets to be used with msm' )
-    packageStartupMessage( 'msmtools requires data.table (>= 1.9.6), msm (>= 1.6) and survival (>= 2.38.0)' )
+    packageStartupMessage( 'msmtools requires data.table (>= 1.9.6), msm (>= 1.6), survival (>= 2.38.0)',
+                           ' and ggplot2 (>=3.3.3)')
     packageStartupMessage( 'For help ?msmtools or vignette( "msmtools" )' )
-    packageStartupMessage( 'If you find bugs, please report them at https://github.com/contefranz/msmtools/issues' )
+    packageStartupMessage( 'If you find a bug, please report it at https://github.com/contefranz/msmtools/issues' )
   }
 }